Refactor Tabs
